requirements.txt
langchain
langchain-openai
python-dotenv
openai
pypdf
sentence_transformers
chromadb
python-telegram-bot
tqdm
pypdfium2
fastapi
python-multipart